+++ assignment-1/my-app/src/app/day-1/Button/page.jsx
@@ -1,0 +1,13 @@
+import Button from '@/app/day-1/components/Button'
+export default function showGreeting(){
+    return(<>
+       <div >
+        <h1>Button Component</h1>
+        <div id="container">
+        <Button text={"button1"} color={"red"} />
+        <Button text={"button2"} color={"yellow"} />
+         <Button text={"button3"} color={"Green"} />
+        </div>
+      </div>
+    </>);
+}
