+++ assignment-1/my-app/src/app/day-1/Counter/page.jsx
@@ -1,0 +1,9 @@
+import Counter from '@/app/day-1/components/Counter'
+export default function showGreeting(){
+    return(<>
+        <div>
+        <h1>Counter Component</h1>
+        <Counter/>
+      </div>
+    </>);
+}
