+++ assignment-1/my-app/src/app/day-1/Greeting/page.jsx
@@ -1,0 +1,8 @@
+import Greeting from '@/app/day-1/components/Greeting'
+export default function showGreeting(){
+    return(<>
+        <div>
+            <Greeting/>
+              </div>
+    </>);
+}
