+++ assignment-1/my-app/src/app/day-1/Weather/page.jsx
@@ -1,0 +1,11 @@
+import Weather from '@/app/day-1/components/Weather'
+export default function showGreeting(){
+    return(<>
+        <div>
+           <h1>Weather Component</h1>
+        <Weather temperature={23} />
+        <Weather temperature={26} />
+        <Weather temperature={5} />
+              </div>
+    </>);
+}
