+++ assignment-1/my-app/src/app/day-1/components/Button.jsx
@@ -2,10 +2,14 @@
 // Style the button using inline styles or CSS classes based on the color prop.
 // Import and render the Button component in the App component with different text and color props.
 
-export default function Button({text,color}){
-    return(
-        <div id="button">
-            <button style={{backgroundColor:color, width: 130, height: 50,fontSize: 30}}>{text}</button>
-        </div>
-    );
+export default function Button({ text, color }) {
+  return (
+    <div id="button">
+      <button
+        style={{ backgroundColor: color, width: 130, height: 50, fontSize: 30 }}
+      >
+        {text}
+      </button>
+    </div>
+  );
 }
