+++ assignment-1/my-app/src/app/day-1/components/Counter.jsx
