+++ assignment-1/my-app/src/app/day-1/components/Greeting.jsx
@@ -1,11 +1,11 @@
 // 1.Create a functional component named Greeting that displays a simple "Hello, Next.js!" message
 //  on the screen.Import and render the Greeting component in the App component.
 
-export default function display(){
-   return (
+export default function display() {
+  return (
     <div>
-        <p id="ass-1">Hello, Next.js</p>
+      <h1>Greeting Component</h1>
+      <p id="ass-1">Hello, Next.js</p>
     </div>
-   );
+  );
 }
-
