+++ assignment-1/my-app/src/app/globals.css
@@ -49,6 +49,18 @@ h1{
 #page{
   margin-bottom: 40px;
 }
+#page button{
+  width: 100px;
+  height: 100px;
+  margin-left: 10px;
+  margin-right: 10px;
+  background-color: black;
+  color: white;
+}
+#page button:hover{
+  background-color:white;
+  color: black;
+}
 
 /* UserCard Component */
 
@@ -93,6 +105,8 @@ h1{
    border: solid 2px black;
    width: 250px;
    height:250px;
+   text-align: left;
+   justify-content: center;
    margin: auto;
 }
